requirements.txt
astroid==2.5.6
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
graphics.py==5.0.1.post1
idna==2.10
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pylint==2.8.3
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
toml==0.10.2
urllib3==1.26.5
wrapt==1.12.1














